Move logic to google login from firebase
