+++ hooks/useFirebaseAuth.tsx
@@ -1,16 +1,11 @@
-import { log } from '@charmverse/core/log';
 import type { FirebaseApp } from 'firebase/app';
 import { initializeApp } from 'firebase/app';
-import type { UserCredential } from 'firebase/auth';
 import {
   getAuth,
-  getRedirectResult,
   GoogleAuthProvider,
   isSignInWithEmailLink,
   sendSignInLinkToEmail,
-  signInWithEmailLink,
-  signInWithPopup,
-  signInWithRedirect
+  signInWithEmailLink
 } from 'firebase/auth';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
@@ -18,11 +13,8 @@ import { useEffect, useState } from 'react';
 import charmClient from 'charmClient';
 import { googleWebClientConfig } from 'config/constants';
 import { useUser } from 'hooks/useUser';
-import type { LoginWithGoogleRequest } from 'lib/google/loginWithGoogle';
 import { getAppUrl } from 'lib/utilities/browser';
-import { ExternalServiceError, InvalidInputError, SystemError } from 'lib/utilities/errors';
-
-import type { AnyIdLogin } from '../components/login/LoginButton';
+import { InvalidInputError } from 'lib/utilities/errors';
 
 import { useLocalStorage } from './useLocalStorage';
 import { useSnackbar } from './useSnackbar';
@@ -31,12 +23,11 @@ export function useFirebaseAuth() {
   const [firebaseApp] = useState<FirebaseApp>(initializeApp(googleWebClientConfig));
   // Google client setup start
   const [provider] = useState(new GoogleAuthProvider());
-  const { user, setUser } = useUser();
+  const { setUser } = useUser();
   const [emailForSignIn, setEmailForSignIn] = useLocalStorage('emailForSignIn', '');
   const router = useRouter();
   const { showMessage } = useSnackbar();
 
-  const [isConnectingGoogle, setIsConnectingGoogle] = useState(false);
   useEffect(() => {
     provider.addScope('email');
     provider.addScope('openid');
@@ -45,109 +36,6 @@ export function useFirebaseAuth() {
       prompt: 'select_account'
     });
   }, []);
-
-  async function getGoogleTokenWithRedirect(): Promise<void> {
-    const auth = getAuth(firebaseApp);
-    auth.languageCode = 'en';
-
-    signInWithRedirect(auth, provider);
-  }
-
-  async function getGoogleRedirectResult(): Promise<LoginWithGoogleRequest> {
-    const auth = getAuth(firebaseApp);
-    auth.languageCode = 'en';
-
-    const result = await getRedirectResult(auth);
-
-    if (!result) {
-      throw new ExternalServiceError(`Could not get credentials from Google`);
-    }
-
-    // This gives you a Google Access Token. You can use it to access the Google API.
-    const credential = GoogleAuthProvider.credentialFromResult(result);
-
-    if (!credential) {
-      throw new ExternalServiceError(`Could not authenticate with Google`);
-    }
-
-    return getGoogleCredentials(result);
-  }
-
-  async function getGoogleToken(): Promise<LoginWithGoogleRequest> {
-    try {
-      const auth = getAuth(firebaseApp);
-      auth.languageCode = 'en';
-
-      const result = await signInWithPopup(auth, provider);
-
-      return getGoogleCredentials(result);
-    } catch (error: any) {
-      if (error instanceof SystemError) {
-        throw error;
-      }
-
-      // Handle Errors here.
-      const errorCode = error?.code;
-      const errorMessage = error?.message;
-      // The email of the user's account used.
-      const email = error?.customData?.email;
-      // The AuthCredential type that was used.
-      const credential = GoogleAuthProvider.credentialFromError(error);
-      log.debug({ errorCode, errorMessage, email, receivedCreds: credential });
-
-      throw error;
-    }
-  }
-
-  function getGoogleCredentials(googleResult: UserCredential): LoginWithGoogleRequest {
-    // This gives you a Google Access Token. You can use it to access the Google API.
-    const credential = GoogleAuthProvider.credentialFromResult(googleResult);
-
-    if (!credential) {
-      throw new ExternalServiceError(`Could not authenticate with Google`);
-    }
-
-    const displayName = googleResult.user.displayName ?? (googleResult.user.email as string);
-
-    return {
-      accessToken: credential?.idToken as string,
-      displayName,
-      avatarUrl: googleResult.user.photoURL as string
-    };
-  }
-
-  async function loginWithGoogle(): Promise<AnyIdLogin> {
-    setIsConnectingGoogle(true);
-    try {
-      const googleToken = await getGoogleToken();
-      const loggedInUser = await charmClient.google.login(googleToken);
-      return { user: loggedInUser, identityType: 'Google', displayName: googleToken.displayName || '' };
-    } finally {
-      setIsConnectingGoogle(false);
-    }
-  }
-
-  async function connectGoogleAccount(): Promise<void> {
-    setIsConnectingGoogle(true);
-    try {
-      const googleToken = await getGoogleToken();
-      const loggedInUser = await charmClient.google.connectAccount(googleToken);
-      setUser(loggedInUser);
-    } finally {
-      setIsConnectingGoogle(false);
-    }
-  }
-
-  async function disconnectGoogleAccount(): Promise<void> {
-    if (!user?.googleAccounts.length) {
-      throw new InvalidInputError('No Google account connected to user');
-    }
-
-    const loggedInUser = await charmClient.google.disconnectAccount({
-      googleAccountEmail: user?.googleAccounts[0].email as string
-    });
-    setUser(loggedInUser);
-  }
 
   async function requestMagicLinkViaFirebase({
     email,
@@ -232,16 +120,10 @@ export function useFirebaseAuth() {
   }
 
   return {
-    loginWithGoogle,
-    connectGoogleAccount,
-    disconnectGoogleAccount,
-    isConnectingGoogle,
     requestMagicLinkViaFirebase,
     validateMagicLink,
     disconnectVerifiedEmailAccount,
     emailForSignIn,
-    setEmailForSignIn,
-    getGoogleRedirectResult,
-    getGoogleTokenWithRedirect
+    setEmailForSignIn
   };
 }
