+++ hooks/usePopupLogin.ts
@@ -55,6 +55,7 @@ export function usePopupLogin<T>() {
   };
 
   return {
-    openPopupLogin
+    openPopupLogin,
+    isPopupLoginOpen: !!popupWindow
   };
 }
